LayoutTests/http/tests/site-isolation/inspector/console/message-from-iframe.html

<script src="/inspector/resources/inspector-test.js"></script>
<script>
    // Utilities for page code.
    function addIFrame(url) {
        const iframe = document.createElement("iframe");
        iframe.src = url;
        document.body.appendChild(iframe);
    }

    function test() {
        // The purpose of these tests are to test that Web Inspector should receive console messages generated from various iframe tree correctly.

        const suite = InspectorTest.createAsyncSuite("MessageFromIFrame");

        // Utilities for test code.
        let callback = () => {};
        WI.consoleManager.addEventListener(WI.ConsoleManager.Event.MessageAdded, (event) => {
            let { target: { type }, messageText } = event.data.message;
            InspectorTest.log(`Got message from ${type} target: ${messageText}`);
            callback();
        });

        function runMessageTest(resolve, expected, func) {
            let seen = 0;

            callback = () => {
                if (++seen === expected)
                    resolve();
            };

            InspectorTest.evaluateInPage(func);
        }

        suite.addTestCase({
            name: "MessageFromIFrame.SameOriginIFrame",
            description: "console message coming from same origin iframe",
            test: (resolve) => runMessageTest(resolve, 3, `addIFrame("resources/console-messages.html")`),
        });

        suite.addTestCase({
            name: "MessageFromIFrame.CrossOriginIFrame",
            description: "console message coming from cross origin iframe",
            test: (resolve) => runMessageTest(resolve, 3, `addIFrame("http://localhost:8000/site-isolation/inspector/console/resources/console-messages.html")`),
        });

        suite.addTestCase({
            name: "MessageFromIFrame.GrandChildSameOriginIFrameInSameOrigin",
            description: "console message coming from same origin grand-child iframe in same origin iframe",
            test: (resolve) => runMessageTest(resolve, 4, `addIFrame("resources/embedded-same-origin.html")`),
        });

        suite.addTestCase({
            name: "MessageFromIFrame.GrandChildSameOriginIFrameInCrossOrigin",
            description: "console message coming from same origin grand-child iframe in cross origin iframe",
            test: (resolve) => runMessageTest(resolve, 4, `addIFrame("http://localhost:8000/site-isolation/inspector/console/resources/embedded-cross-origin.html")`),
        });

        suite.runTestCasesAndFinish();
    }
</script>

<body onload="runTest()"></body>
